requirements.txt
pyinstaller
tkinterdnd2
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
google-cloud-vision
google-cloud-language
google-generativeai
Pillow
moviepy==1.0.3